wan: vae: avoid expensive padding
This CausalConv3d class assumes the worst case scenario of having to
manually pad. However its quite common that that the conditional path
where the cache value outright replaces all padding in the time
dimension. In this case, the manual pad is un-nessacary and burns
both VRAM and bus bandwidth.

Avoid this case by unconditionally moving the spatial padding back
to the symmetric padding supported by torch. Then only pad temporally
if the cat hasn't done it for you already.