mm: lock down the pinning API and its tensors
dont trust the model to be undistrurbed by custom node packs, and
strong ref the tensors as they are pinned. When they are unpinned,
use the pointer and size as pinned to do the free with a warning
on any mismatch.

This is to avoid use-after-free issues, where core comfy pins rams
but the custom node logic walks the model and makes changes. pytorch
can in theory free the pinned tensors when custom logic edits the
models and re-allocation arbitary stuff in the pinned region causing
IO errors much later.